+++ backend/README.md
@@ -1,5 +1,8 @@
-### Backend development
-To start the backend, 
+# Backend development
 
-1. Export *MYSQL_DATABASE*, *MYSQL_USER*, *MYSQL_PASSWORD*, *MYSQL_PORT*, *MYSQL_HOST*
-2. Run `$ npm start`
+* To start the backend,
+    1. Export `MYSQL_DATABASE`, `MYSQL_USER`, `MYSQL_PASSWORD`, `MYSQL_PORT`, `MYSQL_HOST`,
+    2. Run `npm start`,
+    3. Server will be create at <http://127.0.0.1:3000>.
+
+* To gernerate the apidoc, run `npm run apidoc`.
