+++ backend/package.json
@@ -3,13 +3,18 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./bin/www",
+    "apidoc": "apidoc -i routes/ -o apidoc/"
   },
   "dependencies": {
-    "body-parser": "~1.18.2", 
-    "debug": "~2.6.9",
-    "express": "~4.15.5",
-    "morgan": "~1.9.0",
-    "mysql": "~2.15.0"
+    "apidoc": "^0.17.6",
+    "body-parser": "^1.18.2",
+    "debug": "^2.6.9",
+    "express": "^4.15.5",
+    "morgan": "^1.9.0",
+    "mysql": "^2.15.0"
+  },
+  "apidoc": {
+    "title": "Portfolio API"
   }
 }
