+++ script/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "script",
+  "version": "0.0.0",
+  "private": true,
+  "dependencies": {
+    "mysql": "^2.15.0",
+    "moment": "^2.19.1",
+    "when": "^3.7.8",
+    "yahoo-finance": "^0.3.4"
+  }
+}
